requirements.txt
Flask==3.0.0

psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

requests==2.31.0

python-dotenv==1.0.0

pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2

jupyter==1.0.0